+++  my-simple-code/src/game/LoongBgSrv/StateBattlegroundStart.h
@@ -18,7 +18,7 @@ public:
 	StateBattlegroundStart(Battleground* bg);
 	virtual ~StateBattlegroundStart();
 
-	virtual void run(uint32 curTime);
+	virtual void run(int64 curTime);
 
 	virtual BgStateE getState() { return BGSTATE_START; }
 	virtual uint32	getStateTimeLimit() { return 10000; }
