+++  my-simple-code/src/game/LoongBgSrv/scene/Scene.h
@@ -30,7 +30,7 @@ public:
 	void broadMsg(PacketBase& op);
 
 	bool init();
-	void run(uint32 curTime);
+	void run(int64 curTime);
 	void shutdown();
 
 	std::map<int32, BgPlayer*>& getPlayerMgr()
