+++  my-simple-code/src/game/LoongBgSrv/skill/AddSpeedBuf.cc
@@ -9,7 +9,7 @@
 
 #include <assert.h>
 
-AddSpeedBuf::AddSpeedBuf(int16 bufId, uint32 curTime, uint32 bufferTime):
+AddSpeedBuf::AddSpeedBuf(int16 bufId, int64 curTime, int64 bufferTime):
 	Buf(bufId, "提速", curTime, 0, bufferTime)
 {
