+++  my-simple-code/src/game/LoongBgSrv/skill/HunXuanBuf.cc
@@ -9,7 +9,7 @@
 
 #include <assert.h>
 
-HunXuanBuf::HunXuanBuf(int16 bufId, uint32 curTime, uint32 bufferTime):
+HunXuanBuf::HunXuanBuf(int16 bufId, int64 curTime, int64 bufferTime):
 	Buf(bufId, "昏眩", curTime, 0, bufferTime)
 {
